Fix production build validation
