+++ apps/web/app/api/automation-jobs/execute/queue/route.ts
@@ -12,7 +12,7 @@ export const maxDuration = 300;
 
 const logger = createScopedLogger("automation-jobs/execute/queue");
 
-export const POST = handleCallback<z.infer<typeof executeAutomationJobBody>>(
+const queueHandler = handleCallback<z.infer<typeof executeAutomationJobBody>>(
   async (message, metadata) => {
     const parseResult = executeAutomationJobBody.safeParse(message);
     if (!parseResult.success) {
@@ -55,3 +55,5 @@ export const POST = handleCallback<z.infer<typeof executeAutomationJobBody>>(
     }),
   },
 );
+
+export const POST = (request: Request) => queueHandler(request);
