+++ apps/web/components/Celebration.tsx
@@ -1,10 +1,20 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import Confetti from "react-dom-confetti";
+import type { ComponentType } from "react";
+import ReactDomConfetti from "react-dom-confetti";
 import Image from "next/image";
 import { getCelebrationImage } from "@/utils/celebration";
 import { Button } from "@/components/Button";
+
+const Confetti = ReactDomConfetti as unknown as ComponentType<{
+  active: boolean;
+  config: {
+    duration: number;
+    elementCount: number;
+    spread: number;
+  };
+}>;
 
 export function Celebration(props: { message: string }) {
   const [active, setActive] = useState(false);
