+++ apps/web/utils/scripts/lemon.tsx
@@ -1,13 +1,21 @@
 "use client";
 
 import { env } from "@/env";
-import Script from "next/script";
+import NextScript from "next/script";
+import type { ComponentType } from "react";
+
+const ExternalScript = NextScript as unknown as ComponentType<{
+  src: string;
+  defer: boolean;
+  onError: (error: unknown) => void;
+  onLoad: () => void;
+}>;
 
 export function LemonScript() {
   if (!env.NEXT_PUBLIC_LEMON_STORE_ID) return null;
 
   return (
-    <Script
+    <ExternalScript
       src="/vendor/lemon/affiliate.js"
       defer
       onError={(e) => {
